backend/requirements.txt
django
django-rest-framework
djangorestframework-simplejwt
django-cors-headers
whitenoise
django-import-export
python-dotenv
requests